requirements.txt
tkinter>=8.6
Pillow>=9.0.0
requests>=2.27.0
beautifulsoup4>=4.10.0
selenium>=4.1.0
whois>=0.9.3
tldextract>=3.1.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
pandas>=1.3.0
python-whois>=0.8.0
dnspython>=2.2.0
joblib>=1.1.0